feat: #34 GET /invitation を /admin/invitation に移動、招待APIを丸ごと管理APIへ
- 全バックエンド: GET /invitation → GET /admin/invitation
  - Go: invitation_handler.go 削除、AdminInvitationHandler に Index 追加
  - Python: invitations.py 削除、admin_invitations.py に index 追加
  - TypeScript: invitations.ts 削除、adminInvitations.ts に current 追加
  - PHP: Api/InvitationController.php 削除、Admin/InvitationController に index 追加
- 旧テストファイル削除、Admin 用テストに Index テスト追加（全バックエンド）
- openapi.yml: /invitation → /admin/invitation、invitation タグ廃止
- フロントエンド: getInvitation エンドポイントを /admin/invitation に変更
- E2E テスト: モック URL を /admin/invitation に更新
