+++ backends/go-gin/internal/handler/admin_invitation_handler.go
@@ -15,6 +15,21 @@ func NewAdminInvitationHandler(svc *uinvitation.Interactor) *AdminInvitationHand
 	return &AdminInvitationHandler{svc: svc}
 }
 
+// GET /api/admin/invitation
+func (h *AdminInvitationHandler) Index(c *gin.Context) {
+	result, err := h.svc.Current()
+	if err != nil {
+		_ = c.Error(err)
+		return
+	}
+	c.JSON(http.StatusOK, gin.H{
+		"found":       true,
+		"url":         result.URL,
+		"display_url": result.DisplayURL,
+		"token":       result.Token,
+	})
+}
+
 // GET /api/admin/invitation/issue
 func (h *AdminInvitationHandler) Issue(c *gin.Context) {
 	result, err := h.svc.Issue()
