+++ backends/python-fastapi/app/routers/admin_invitations.py
@@ -5,6 +5,12 @@ from app.usecase.invitation.interactor import InvitationInteractor
 router = APIRouter()
 
 
+@router.get("/admin/invitation")
+def index(interactor: InvitationInteractor = Depends(get_invitation_interactor)):
+    result = interactor.current()
+    return {"found": True, "url": result.url, "display_url": result.display_url, "token": result.token}
+
+
 @router.get("/admin/invitation/issue")
 def issue(interactor: InvitationInteractor = Depends(get_invitation_interactor)):
     result = interactor.issue()
