+++ backends/python-fastapi/tests/test_admin_invitations.py
@@ -3,6 +3,16 @@
 from tests.conftest import make_invitation
 
 
+class TestAdminInvitationIndex:
+    def test_現在の招待URLが取得できる(self, client, db_session):
+        inv = make_invitation(db_session, token="current-token")
+        res = client.get("/api/admin/invitation")
+        assert res.status_code == 200
+        data = res.json()
+        assert data["token"] == inv.token
+        assert "url" in data
+
+
 class TestAdminInvitationIssue:
     def test_招待URLが発行できる(self, client):
         res = client.get("/api/admin/invitation/issue")
