+++ backends/ts-hono/src/app.ts
@@ -5,7 +5,6 @@ import { config } from "./config.js";
 import authRoutes, { oauthApp } from "./routes/auth.js";
 import clientRoutes from "./routes/clients.js";
 import staffRoutes from "./routes/staffs.js";
-import invitationRoutes from "./routes/invitations.js";
 import adminInvitationRoutes from "./routes/adminInvitations.js";
 import gateRoutes from "./routes/gate.js";
 import notificationRoutes from "./routes/notifications.js";
@@ -34,7 +33,6 @@ export function createApp(): Hono {
   api.route("/", authRoutes);
   api.route("/", clientRoutes);
   api.route("/", staffRoutes);
-  api.route("/", invitationRoutes);
   api.route("/admin", adminInvitationRoutes);
   api.route("/", gateRoutes);
   api.route("/", notificationRoutes);
